Switch to Maven 3.2.1 and Android plugin 3.8.2; enable incremental dexing
Some notes: The android-maven-plugin 3.8.2 requires Maven 3.1.1+ and vice versa.

The android-maven-plugin 3.8.2 has the advantage of supporting incremental
dexing, which reduces build times. Older versions only suport pre-dexing, which
also helps with build times, but not as much.

Compiling with android-maven-plugin 3.8.2 seems to result in an
ActionBarSherlock NPE when running the resulting app on pre-4.0 devices. Hive
Android does not support those anyway, so it is not a problem for us.

Maven 3.1.1 or higher is also a requirement for using the Maven Android SDK
Deployer.